Crate move_idl_types

Rust types for the Move IDL specification.

Macros

ident_str

ident_str! is a compile-time validated macro that constructs a &'static IdentStr from a const &'static str.

Structs

AccountAddress

A struct that represents an account address.

AccountAddressData

Wrapper around AccountAddress which is serialized as a string.

IDLArgument

IDLError

IDL error.

IDLErrorMapping

IDL error mapping.

IDLField

A struct.

IDLModule

Simplified ABI for a Move module.

IDLPackage

A set of modules.

IDLScriptFunction

IDLStruct

A struct.

IDLStructType

A struct with types.

IDLTypeParam

A struct type parameter.

Identifier

An owned identifier.

ModuleId

Represents the initial key into global storage where we first index by the address, and then the struct tag

ModuleIdData

Wrapper around ModuleId which is serialized as a string.

StructTag

StructTagData

Wrapper around StructTag which is serialized as a string.

Enums

IDLAbility

An Ability classifies what operations are permitted for a given type

IDLType

A type represented in the IDL.

Functions

parse_module_id

Parses a module ID.

parse_struct_tag

Parses a struct tag.